controller修正
